Feat: 라운드 결과 출력 기능 구현

+++ src/main/java/racingcar/controller/MainController.java
@@ -46,7 +46,7 @@ public class MainController {
 
     private TotalRound createTotalRound() {
         return readUserInput(() -> {
-            long totalRound = inputView.readTotalRound();
+            int totalRound = inputView.readTotalRound();
             return TotalRound.from(totalRound);
         });
     }
