+++ src/main/java/racingcar/domain/TotalRound.java
@@ -3,18 +3,18 @@ package racingcar.domain;
 import racingcar.utils.TotalRoundValidator;
 
 public class TotalRound {
-    private final long totalRound;
+    private final int totalRound;
 
-    private TotalRound(long totalRound) {
+    private TotalRound(int totalRound) {
         this.totalRound = totalRound;
     }
 
-    public static TotalRound from(long totalRound) {
+    public static TotalRound from(int totalRound) {
         TotalRoundValidator.validatePositive(totalRound);
         return new TotalRound(totalRound);
     }
 
-    public long getTotalRound() {
+    public int getTotalRound() {
         return totalRound;
     }
 }
